+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next version
+- Add option to install Metrics Server
 - Fix "Inspect node pools" macro when using managed node groups
 - Remove macro `Run Kubectl command` (natively supported in DSS 10.0.6)
 - Support spot instances in node groups