+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix "Inspect node pools" macro when using managed node groups
 - Support tagging nodes
 - Remove macro `Run Kubectl command` (natively supported in DSS 10.0.6)
+- Support spot instances in node groups
 
 ## Version 1.0.7 - Bugfix release
 - Add capability to assume IAM role on all cluster operation